backend/pyproject.toml
[tool.poetry]
name = "bahamm-backend"
version = "0.1.0"
description = ""
authors = ["Shayan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.11"
sqlalchemy = "^2.0.38"
pydantic = "^2.10.6"
email-validator = "^2.2.0"
uvicorn = "^0.34.0"
passlib = "1.7.4"
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
pydantic-settings = "^2.8.1"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
